Updated submodules and WIP api
