+++ indexer/indexer/api/main.py
@@ -5,10 +5,9 @@ from fastapi.responses import JSONResponse
 from fastapi.exceptions import RequestValidationError
 from starlette.exceptions import HTTPException as StarletteHTTPException
 
-from indexer.core import crud
 from indexer.api.api_v0.main import router as router_v0
-from indexer.api.api_v1.main import router as router_v1
 
+from indexer.core import exceptions
 from indexer.core.settings import Settings
 
 settings = Settings()
@@ -23,14 +22,13 @@ app = FastAPI(title="TON Index",
               description=description,
               version='0.1.0',
               root_path=settings.api_root_path,
-              # openapi_url=settings.api_root_path + '/openapi.json',
               docs_url='/')
 
 @app.exception_handler(StarletteHTTPException)
 async def http_exception_handler(request, exc):
     return JSONResponse({'error' : str(exc.detail)}, status_code=exc.status_code)
 
-@app.exception_handler(crud.DataNotFound)
+@app.exception_handler(exceptions.DataNotFound)
 async def tonlib_wront_result_exception_handler(request, exc):
     return JSONResponse({'error' : str(exc)}, status_code=status.HTTP_404_NOT_FOUND)
 
@@ -48,6 +46,4 @@ def generic_exception_handler(request, exc):
 def startup():
     logger.info('Service started successfully')
 
-app.include_router(router_v1, prefix='/v1')
-app.include_router(router_v0, prefix='/v0', include_in_schema=False, deprecated=True)
-app.include_router(router_v0, prefix='', include_in_schema=True, deprecated=True)
+app.include_router(router_v0, prefix='/v0', include_in_schema=True, deprecated=False)
